requirements.txt
appdirs==1.4.3
asgiref==3.5.2
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.0.12
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.12.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.4.1
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
pipenv==11.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
pytz==2022.2.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==27.9
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
